Improve button on top
